Correct number of tasks

+++ openMP/main.cpp
@@ -55,23 +55,23 @@ int main(int argc, char *argv[]) {
 
     // start the actual benchmark workload
     for(int i = 0; i < turns; i++){
-	#pragma omp parallel
+	    #pragma omp parallel
         {
-	    #pragma omp single
+	        #pragma omp single
             {
-                int nrTasks = (int) std::ceil(VECTOR_SIZE / taskSize);
-		for (int k = 0; k < nrTasks - 1; k++) {
-		    #pragma omp task
-		    {
-			for(int j = k * taskSize; j < (k + 1) * taskSize; j++) {
-                            vectors[(i + 1) % 2][j] = fabsf(std::sin(vectors[i % 2][j])) * 10;
-			}
-		    }
-		}
-		for(int j = (nrTasks - 1) * taskSize; j < VECTOR_SIZE; j++) {
+                int nrTasks = (int) std::ceil(VECTOR_SIZE / (float) taskSize);
+		        for (int k = 0; k < (nrTasks - 1); k++) {
+		            #pragma omp task
+		            {
+			            for(int j = k * taskSize; j < (k + 1) * taskSize; j++) {
+			                vectors[(i + 1) % 2][j] = fabsf(std::sin(vectors[i % 2][j])) * 10;
+			            }
+		            }
+		        }
+		        for(int j = (nrTasks - 1) * taskSize; j < VECTOR_SIZE; j++) {
                     vectors[(i + 1) % 2][j] = fabsf(std::sin(vectors[i % 2][j])) * 10;
-		}
-	    }
+		        }
+            }
         }
     }
